requirements.txt
# Discord Music Bot Requirements

# Discord API
discord.py>=2.3.1

# Environment Variables
python-dotenv>=1.0.0

# Audio Processing
PyNaCl>=1.5.0  # Required for voice support

# Media Downloading
yt-dlp>=2023.3.4

# FFmpeg is also required but is installed system-wide, not via pip
